combined a new version to put dataframe to datawarehouse whether insert or drop
